Add UI instructions and fix calculation bug
Fixed bug in measuring objects from photos. Photos rescaled in different context on screen. But a constant exact screen size representation of a photo is required, regardless of the page it is displayed on. This is because the app is calibrated using a certain photo representation scale on the device`s screen.